CA-186157: Extract tapdisk stats to separate header
Exported tapdisk-metrics.h is impossible to include,
since it contains dependencies to other blktap headers
which for now remains private (tapdisk.h, tapdisk-utils.h).

This patch moves tapdisk stats struct to a header
that does not contains other dependencies,
so that it can be included by other components
that need access to tapdisk statistics structure
