+++ include/Makefile.am
@@ -21,7 +21,7 @@ blktap_HEADERS += util.h
 blktap_HEADERS += payload.h
 blktap_HEADERS += kpr_util.h
 blktap_HEADERS += thin_log.h
-blktap_HEADERS += ../drivers/tapdisk-metrics.h
+blktap_HEADERS += ../drivers/tapdisk-metrics-stats.h
 
 noinst_HEADERS  = blktap.h
 noinst_HEADERS += compiler.h
